Make babel-gatsby-preset nested presets configurable
Hi there, 

at the moment it's not possible to configure the nested presets or plugins in the gatsby babel preset. Whenever we need to adjust a nested preset like @babel/preset-react we need to build a copied babelrc of the standard config. With this PR it's possible to put "suboptions" into the options object ex: { "@babel/preset-react": { "useBuiltIns": false } }. This way we can also use setBabelPreset action to adjust the config easily from a plugin.